Added constructor to personAgent so the add button now actually creates people.  Still need to decide on how to set occupation for market and restaurant.  Next I will add the gui options for shift time and weekends vs weekdays. Issue #8.

+++ source/gui/ControlPanel.java
@@ -45,7 +45,7 @@ public class ControlPanel extends JTabbedPane {
 		this.setSelectedComponent(currentBuildingPanel);
 	}
 
-	public void addPerson(String name, int money, String occupation) 
+	public void addPerson(String name, double money, String occupation) 
 	{
 		currentPersonPanel.addPerson(name);
 		// TODO this is where we should actually create the new Person agent.  We should discuss how we want to handle the agents.
